pi-plant-watcher/sensor/requirements.txt
azure-iot-device==2.11.0